Some more errors from online Doctests cleared

+++ pandas/core/algorithms.py
@@ -346,7 +346,7 @@ def unique(values):
     array([2, 1, 3])
 
     >>> pd.unique(pd.Series([2] + [1] * 5))
-    array([2, 1], dtype=int64)
+    array([2, 1])
 
     >>> pd.unique(pd.Series([pd.Timestamp("20160101"), pd.Timestamp("20160101")]))
     array(['2016-01-01T00:00:00.000000000'], dtype='datetime64[ns]')
@@ -714,7 +714,7 @@ def factorize(
 
     >>> codes, uniques = pd.factorize(values, use_na_sentinel=False)
     >>> codes
-    array([0, 1, 0, 2], dtype=int64)
+    array([0, 1, 0, 2])
     >>> uniques
     array([ 1.,  2., nan])
     """
